+++ packages/core/src/connectivity/scp-cf/identity-service.ts
@@ -1,17 +1,23 @@
+import * as xssec from '@sap/xssec';
 import { DestinationOptions } from './destination';
 import { getRawXsuaaServiceCredentials } from './environment-accessor';
-import * as xssec from '@sap/xssec';
 
 /**
  * Make a token exchange from IAS token to XSUAA token.
  * @param options - Configuration for how to retrieve destinations from the destination service.
  * @returns Exchanged token.
  */
-export async function exchangeToken(options: DestinationOptions): Promise<string> {
+export async function exchangeToken(
+  options: DestinationOptions
+): Promise<string> {
   const xsuaaServiceCredentials = getRawXsuaaServiceCredentials();
-  return await new Promise((resolve: (p: string) => void, reject) => {
-    xssec.createSecurityContext(options.userJwt, xsuaaServiceCredentials,
-      (err: Error, context, tokenInfo) => (err ? reject(err) : resolve(tokenInfo.getTokenValue())));
+  return new Promise((resolve: (p: string) => void, reject) => {
+    xssec.createSecurityContext(
+      options.userJwt,
+      xsuaaServiceCredentials,
+      (err: Error, context, tokenInfo) =>
+        err ? reject(err) : resolve(tokenInfo.getTokenValue())
+    );
   });
 }
 
@@ -20,6 +26,6 @@ export async function exchangeToken(options: DestinationOptions): Promise<string
  * @param options - Configuration for how to retrieve destinations from the destination service.
  * @returns A boolean value, that indicates whether the token exchange should be applied.
  */
-export function isTokenExchangeEnabled(options: DestinationOptions): boolean{
+export function isTokenExchangeEnabled(options: DestinationOptions): boolean {
   return options.iasToXsuaaTokenExchange !== false && !!options.userJwt;
 }